Split flow.go into three focused files
- flow.go (185 lines): types, constants, Flow struct, core dispatch
- flow_states.go (377 lines): all state transition implementations
- flow_render.go (157 lines): lifecycle helpers and state rendering

Pure refactor — no behavior changes. All tests pass.

Closes #10
